refactor: extract anthropic beta handler into transform chain and result dispatch.
