+++ internal/server/anthropic_beta.go
@@ -2,20 +2,16 @@ package server
 
 import (
 	"context"
-	"fmt"
 	"net/http"
 
 	"github.com/anthropics/anthropic-sdk-go"
 	anthropicstream "github.com/anthropics/anthropic-sdk-go/packages/ssestream"
 	"github.com/gin-gonic/gin"
-	"github.com/openai/openai-go/v3"
 	"github.com/openai/openai-go/v3/responses"
 	"github.com/sirupsen/logrus"
 	"github.com/tingly-dev/tingly-box/internal/protocol"
 	"github.com/tingly-dev/tingly-box/internal/protocol/nonstream"
-	"github.com/tingly-dev/tingly-box/internal/protocol/request"
 	"github.com/tingly-dev/tingly-box/internal/protocol/stream"
-	"github.com/tingly-dev/tingly-box/internal/protocol/transform"
 	serverguardrails "github.com/tingly-dev/tingly-box/internal/server/guardrails"
 	"github.com/tingly-dev/tingly-box/internal/toolinterceptor"
 	"github.com/tingly-dev/tingly-box/internal/typ"
@@ -75,314 +71,41 @@ func (s *Server) AnthropicMessagesV1Beta(c *gin.Context, req protocol.AnthropicB
 	// Check provider's API style to decide which path to take
 	apiStyle := provider.APIStyle
 
+	target := protocol.APIAnthropicBeta
 	switch apiStyle {
 	case protocol.APIStyleAnthropic:
-		// Use direct Anthropic SDK call
-		wrapper := s.clientPool.GetAnthropicClient(provider, string(req.BetaMessageNewParams.Model))
-		fc := NewForwardContext(c.Request.Context(), provider)
-
-		finalCtx, err := s.transformAnthropicBeta(c, req, transform.TargetAPIStyleAnthropicBeta, provider, isStreaming, recorder, scenarioType)
-		if err != nil {
-			c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-			if recorder != nil {
-				recorder.RecordError(err)
-			}
-			return
-		}
-		// Get final transformed request
-		transformedReq := finalCtx.Request.(*anthropic.BetaMessageNewParams)
-
-		if isStreaming {
-			// Handle streaming request with request context for proper cancellation
-			streamResp, cancel, err := ForwardAnthropicV1BetaStream(fc, wrapper, *transformedReq)
-			if cancel != nil {
-				defer cancel()
-			}
-			if err != nil {
-				s.trackUsageFromContext(c, 0, 0, err)
-				stream.SendStreamingError(c, err)
-				if recorder != nil {
-					recorder.RecordError(err)
-				}
-				return
-			}
-			// Handle the streaming response
-			s.handleAnthropicStreamResponseV1Beta(c, req.BetaMessageNewParams, streamResp, proxyModel, actualModel, provider, recorder)
-		} else {
-			// Handle non-streaming request
-			anthropicResp, cancel, err := ForwardAnthropicV1Beta(fc, wrapper, req.BetaMessageNewParams)
-			if cancel != nil {
-				defer cancel()
-			}
-			if err != nil {
-				s.trackUsageFromContext(c, 0, 0, err)
-				stream.SendForwardingError(c, err)
-				if recorder != nil {
-					recorder.RecordError(err)
-				}
-				return
-			}
-
-			// Track usage from response
-			inputTokens := int(anthropicResp.Usage.InputTokens)
-			outputTokens := int(anthropicResp.Usage.OutputTokens)
-			cacheTokens := int(anthropicResp.Usage.CacheReadInputTokens)
-			usage := protocol.NewTokenUsageWithCache(inputTokens, outputTokens, cacheTokens)
-			s.trackUsageWithTokenUsage(c, usage, nil)
-
-			// Update affinity entry with message ID
-			s.updateAffinityMessageID(c, rule, string(anthropicResp.ID))
-
-			// FIXME: now we use req model as resp model
-			anthropicResp.Model = anthropic.Model(proxyModel)
-
-			session := s.guardrailsSessionFromContext(c, actualModel, provider)
-			messageHistory := serverguardrails.MessagesFromAnthropicV1Beta(req.System, req.Messages)
-			blocked := s.applyGuardrailsToAnthropicV1BetaNonStreamResponse(c, session, messageHistory, anthropicResp)
-			if !blocked {
-				s.restoreGuardrailsCredentialAliasesV1BetaResponse(c, anthropicResp)
-			}
-
-			// Record response if scenario recording is enabled
-			if recorder != nil {
-				recorder.SetAssembledResponse(anthropicResp)
-				recorder.RecordResponse(provider, actualModel)
-			}
-			c.JSON(http.StatusOK, anthropicResp)
-		}
-		return
-
+		target = protocol.APIAnthropicBeta
 	case protocol.APIStyleGoogle:
-		// Convert Anthropic beta request to Google format
-		model, googleReq, cfg := request.ConvertAnthropicBetaToGoogleRequest(&req.BetaMessageNewParams, 0)
-
-		if isStreaming {
-			// Create streaming request with request context for proper cancellation
-			wrapper := s.clientPool.GetGoogleClient(provider, model)
-			fc := NewForwardContext(c.Request.Context(), provider)
-			streamResp, cancel, err := ForwardGoogleStream(fc, wrapper, model, googleReq, cfg)
-			if cancel != nil {
-				defer cancel()
-			}
-			if err != nil {
-				stream.SendStreamingError(c, err)
-				if recorder != nil {
-					recorder.RecordError(err)
-				}
-				return
-			}
-
-			// Handle the streaming response
-			usage, err := stream.HandleGoogleToAnthropicBetaStreamResponse(c, streamResp, proxyModel)
-			if err != nil {
-				s.trackUsageWithTokenUsage(c, usage, err)
-				stream.SendInternalError(c, err.Error())
-				if recorder != nil {
-					recorder.RecordError(err)
-				}
-				return
-			}
-
-			// Track usage from stream handler
-			s.trackUsageWithTokenUsage(c, usage, nil)
-
-		} else {
-			// Handle non-streaming request
-			wrapper := s.clientPool.GetGoogleClient(provider, model)
-			fc := NewForwardContext(nil, provider)
-			resp, _, err := ForwardGoogle(fc, wrapper, model, googleReq, cfg)
-			if err != nil {
-				stream.SendForwardingError(c, err)
-				if recorder != nil {
-					recorder.RecordError(err)
-				}
-				return
-			}
-
-			// Convert Google response to Anthropic beta format
-			anthropicResp := nonstream.ConvertGoogleToAnthropicBetaResponse(resp, proxyModel)
-
-			// Track usage from response
-			inputTokens := 0
-			outputTokens := 0
-			cacheTokens := 0
-			if resp.UsageMetadata != nil {
-				inputTokens = int(resp.UsageMetadata.PromptTokenCount)
-				outputTokens = int(resp.UsageMetadata.CandidatesTokenCount)
-				cacheTokens = int(resp.UsageMetadata.CachedContentTokenCount)
-			}
-			usage := protocol.NewTokenUsageWithCache(inputTokens, outputTokens, cacheTokens)
-			s.trackUsageWithTokenUsage(c, usage, nil)
-
-			// Update affinity entry with message ID
-			s.updateAffinityMessageID(c, rule, string(anthropicResp.ID))
-
-			// Record response if scenario recording is enabled
-			if recorder != nil {
-				recorder.SetAssembledResponse(anthropicResp)
-				recorder.RecordResponse(provider, actualModel)
-			}
-			c.JSON(http.StatusOK, anthropicResp)
-		}
-
+		target = protocol.APIGoogle
 	case protocol.APIStyleOpenAI:
 		// Check if model prefers Responses API (for models like Codex)
 		// This is used for ChatGPT backend API which only supports Responses API
 		preferredEndpoint := s.GetPreferredEndpointForModel(provider, actualModel)
 		logrus.Debugf("[AnthropicV1] Probe cache preferred endpoint for model=%s: %s", actualModel, preferredEndpoint)
 		useResponsesAPI := preferredEndpoint == "responses"
-
 		if useResponsesAPI {
-			// Use Responses API path with Transform Chain
-			// Set the rule and provider in context so middleware can use the same rule
-			if rule != nil {
-				c.Set("rule", rule)
-			}
-
-			// Set provider UUID in context
-			c.Set("provider", provider.UUID)
-			c.Set("model", actualModel)
-
-			// Set context flag to indicate original request was v1 format
-			// The ChatGPT backend streaming handler will use this to send responses in v1 format
-			c.Set("original_request_format", "v1")
-
-			logrus.Debugf("[Anthropic Beta] Using Transform Chain for Responses API for model=%s", actualModel)
-
-			finalCtx, err := s.transformAnthropicBeta(c, req, transform.TargetAPIStyleOpenAIResponses, provider, isStreaming, recorder, scenarioType)
-			if err != nil {
-				c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-				if recorder != nil {
-					recorder.RecordError(err)
-				}
-				return
-			}
-
-			// Get final transformed request
-			transformedReq := finalCtx.Request.(*responses.ResponseNewParams)
-
-			if isStreaming {
-				req.Stream = true
-				s.handleAnthropicV1BetaViaResponsesAPIStreaming(c, req, proxyModel, actualModel, provider, *transformedReq)
-			} else {
-				if provider.APIBase == protocol.CodexAPIBase {
-					req.Stream = true
-					s.handleAnthropicV1BetaViaResponsesAPIAssembly(c, req, proxyModel, actualModel, provider, *transformedReq)
-				} else {
-					s.handleAnthropicV1BetaViaResponsesAPINonStreaming(c, req, proxyModel, actualModel, provider, *transformedReq)
-				}
-			}
-			return
-		}
-		// Set the rule and provider in context so middleware can use the same rule
-		if rule != nil {
-			c.Set("rule", rule)
-		}
-
-		// Set provider UUID in context
-		c.Set("provider", provider.UUID)
-		c.Set("model", actualModel)
-
-		finalCtx, err := s.transformAnthropicBeta(c, req, transform.TargetAPIStyleOpenAIChat, provider, isStreaming, recorder, scenarioType)
-		if err != nil {
-			c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-			if recorder != nil {
-				recorder.RecordError(err)
-			}
-			return
-		}
-
-		// Get final transformed request
-		transformedReq := finalCtx.Request.(*openai.ChatCompletionNewParams)
-
-		// Clean up temporary fields (e.g., x_thinking)
-		request.CleanupOpenaiFields(transformedReq)
-
-		// Use OpenAI Chat Completions path
-		if isStreaming {
-			// Set up stream recorder
-			streamRec := newStreamRecorder(recorder)
-			if streamRec != nil {
-				streamRec.SetupStreamRecorderInContext(c, "stream_event_recorder")
-			}
-
-			// Create streaming request with request context for proper cancellation
-			wrapper := s.clientPool.GetOpenAIClient(provider, transformedReq.Model)
-			fc := NewForwardContext(c.Request.Context(), provider)
-			streamResp, cancel, err := ForwardOpenAIChatStream(fc, wrapper, transformedReq)
-			if cancel != nil {
-				defer cancel()
-			}
-			if err != nil {
-				stream.SendStreamingError(c, err)
-				if streamRec != nil {
-					streamRec.RecordError(err)
-				}
-				return
-			}
-
-			// Handle the streaming response
-			usage, err := stream.HandleOpenAIToAnthropicBetaStream(c, transformedReq, streamResp, proxyModel)
-			if err != nil {
-				s.trackUsageWithTokenUsage(c, usage, err)
-				stream.SendInternalError(c, err.Error())
-				if streamRec != nil {
-					streamRec.RecordError(err)
-				}
-				return
-			}
-
-			// Track usage from stream handler
-			s.trackUsageWithTokenUsage(c, usage, nil)
-
-			// Finish recording and assemble response
-			if streamRec != nil {
-				streamRec.Finish(proxyModel, usage.InputTokens, usage.OutputTokens)
-				streamRec.RecordResponse(provider, actualModel)
-			}
-
+			target = protocol.APIOpenAIResponses
 		} else {
-			wrapper := s.clientPool.GetOpenAIClient(provider, transformedReq.Model)
-			fc := NewForwardContext(nil, provider)
-			resp, _, err := ForwardOpenAIChat(fc, wrapper, transformedReq)
-			if err != nil {
-				stream.SendForwardingError(c, err)
-				if recorder != nil {
-					recorder.RecordError(err)
-				}
-				return
-			}
-			// Convert OpenAI response back to Anthropic beta format
-			anthropicResp := nonstream.ConvertOpenAIToAnthropicBetaResponse(resp, proxyModel)
-
-			// Track usage from response
-			inputTokens := int(resp.Usage.PromptTokens)
-			outputTokens := int(resp.Usage.CompletionTokens)
-			cacheTokens := int(resp.Usage.PromptTokensDetails.CachedTokens)
-			usage := protocol.NewTokenUsageWithCache(inputTokens, outputTokens, cacheTokens)
-			s.trackUsageWithTokenUsage(c, usage, nil)
-
-			// Update affinity entry with message ID
-			s.updateAffinityMessageID(c, rule, string(anthropicResp.ID))
-
-			// Record response if scenario recording is enabled
-			if recorder != nil {
-				recorder.SetAssembledResponse(anthropicResp)
-				recorder.RecordResponse(provider, actualModel)
-			}
-			c.JSON(http.StatusOK, anthropicResp)
-		}
-	default:
-		c.JSON(http.StatusBadRequest, "tingly-box: invalid api style")
-		if recorder != nil {
-			recorder.RecordError(fmt.Errorf("invalid api style: %s", apiStyle))
+			target = protocol.APIOpenAIChat
 		}
 	}
+
+	reqCtx, err := s.transformAnthropicBeta(c, req, target, provider, isStreaming, recorder, scenarioType)
+	if err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		if recorder != nil {
+			recorder.RecordError(err)
+		}
+		return
+	}
+
+	s.dispatchAnthropicBetaChainResult(c, reqCtx, rule, provider, isStreaming, recorder, actualModel, proxyModel)
 }
 
 // handleAnthropicStreamResponseV1Beta processes the Anthropic beta streaming response and sends it to the client
 func (s *Server) handleAnthropicStreamResponseV1Beta(c *gin.Context, req *anthropic.BetaMessageNewParams, streamResp *anthropicstream.Stream[anthropic.BetaRawMessageStreamEventUnion], respModel string, provider *typ.Provider, recorder *ProtocolRecorder) {
 	hc := protocol.NewHandleContext(c, respModel)
+	actualModel := string(req.Model)
 
 	// Record TTFT when the first streaming chunk arrives
 	firstTokenRecorded := false
